fix: move app.js file
