+++ src/app.js
@@ -4,8 +4,8 @@ import React from 'react'
 import { HashRouter, Route, Routes } from 'react-router-dom'
 import { QueryParamProvider } from 'use-query-params'
 import { ReactRouter6Adapter } from 'use-query-params/adapters/react-router-6'
-import { AppProvider, UserProvider } from '../context/index.js'
-import { DataPage, EditPage, EditItem, AddItem } from '../pages/index.js'
+import { AppProvider, UserProvider } from './context/index.js'
+import { DataPage, EditPage, EditItem, AddItem } from './pages/index.js'
 
 const App = ({ router: Router }) => (
     <>
