--- src/app/app.test.js
@@ -1,9 +1,0 @@
-// import { shallow } from 'enzyme'
-// import React from 'react'
-// import { App } from './app.js'
-
-describe('<App>', () => {
-    it('renders without errors', () => {
-        // shallow(<App />)
-    })
-})
