+++ src/index.js
@@ -1,4 +1,4 @@
 import { AppWrapper } from './app.js'
-import '../locales/index.js'
+import './locales/index.js'
 
 export default AppWrapper
